fix: extract more constants
