+++ tests/utils/constants.py
@@ -1,9 +1,11 @@
 from aidial_analytics_realtime.time import parse_time
 
-DEFAULT_RESPONSE_TIME = "2023-08-16T19:42:39.997"
+DEFAULT_RESPONSE_TIME = "2001-01-01T01:01:01.001"
 DEFAULT_TIMESTAMP = parse_time(DEFAULT_RESPONSE_TIME)
 
-DEFAULT_PROJECT_ID = "PROJECT-KEY"
-DEFAULT_RESPONSE_ID = "chatcmpl-1"
-DEFAULT_CHAT_ID = "chat-1"
+DEFAULT_PROJECT_ID = "default-project-id"
+DEFAULT_RESPONSE_ID = "default-response-id"
+DEFAULT_CHAT_ID = "default-chat-id"
+DEFAULT_USER_ID = "default-user-id"
+DEFAULT_USER_TITLE = "default-user-title"
 DEFAULT_TOPIC = "ping?\n\npong"
