+++ tests/utils/influx.py
@@ -8,6 +8,8 @@ from tests.utils.constants import (
     DEFAULT_RESPONSE_ID,
     DEFAULT_TIMESTAMP,
     DEFAULT_TOPIC,
+    DEFAULT_USER_ID,
+    DEFAULT_USER_TITLE,
 )
 
 
@@ -15,10 +17,10 @@ def create_rate_point(
     *,
     deployment: str = "gpt-4",
     project_id: str = DEFAULT_PROJECT_ID,
-    response_id: str = "rate-response-id",
+    response_id: str = DEFAULT_RESPONSE_ID,
     chat_id: str | None = DEFAULT_CHAT_ID,
-    user_hash: str | None = "test-user-id",
-    title: str | None = "test-user-title",
+    user_hash: str | None = DEFAULT_USER_ID,
+    title: str | None = DEFAULT_USER_TITLE,
     timestamp: datetime = DEFAULT_TIMESTAMP,
     like_count: int = 1,
     dislike_count: int = 0,
@@ -43,17 +45,17 @@ def create_point(
     execution_path: str | None = "app1/app2",
     model: str = "gpt-4",
     deployment: str = "gpt-4",
-    trace_id: str | None = "test-trace-id",
-    core_span_id: str | None = "test-core-span-id",
-    core_parent_span_id: str | None = "core-parent-span-id",
+    trace_id: str | None = "default-trace-id",
+    core_span_id: str | None = "default-core-span-id",
+    core_parent_span_id: str | None = "default-core-parent-span-id",
     project_id: str = DEFAULT_PROJECT_ID,
     response_id: str = DEFAULT_RESPONSE_ID,
-    user_hash: str | None = "test-user-id",
+    user_hash: str | None = DEFAULT_USER_ID,
     chat_id: str = DEFAULT_CHAT_ID,
     number_request_messages: int = 2,
     upstream: str | None = "http://upstream.domain.com/endpoint",
     topic: str | None = DEFAULT_TOPIC,
-    title: str | None = "test-user-title",
+    title: str | None = DEFAULT_USER_TITLE,
     timestamp: datetime = DEFAULT_TIMESTAMP,
     prompt_tokens: int = 0,
     completion_tokens: int = 0,
